Fixed registration.
